newesrererert

+++ libft/ft_strlcpy.c
@@ -6,14 +6,14 @@
 /*   By: adhaka <adhaka@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/10 18:14:25 by adhaka            #+#    #+#             */
-/*   Updated: 2022/12/15 04:02:33 by adhaka           ###   ########.fr       */
+/*   Updated: 2022/12/26 16:27:42 by adhaka           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdio.h>
 #include "libft.h"
 
-int	ft_strlcpy(char *dst, const char *src, size_t size)
+size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
 	size_t		a;
