+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: adhaka <adhaka@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/10 17:48:11 by adhaka            #+#    #+#             */
-/*   Updated: 2022/11/23 18:44:50 by adhaka           ###   ########.fr       */
+/*   Updated: 2022/12/26 16:26:47 by adhaka           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-// #include <stdio.h>
+#include <stdio.h>
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
